5 commit ajuste final

+++ PracticaTeoremaPitagoras/src/clases/TrianguloRectangulo.java
@@ -51,7 +51,7 @@ public class TrianguloRectangulo {
 	 */
 	public double hipotenusa() {
 		double hipotenusa = Math.sqrt((Math.pow(this.a, 2))+(Math.pow(this.b, 2)));
-		double hipotenusa2 = Math.round(hipotenusa*100)/100;
+		double hipotenusa2 = (double) Math.round(hipotenusa*100)/100;
 		return hipotenusa2;
 	}
